Move helper channel reducers
