+++ packages/wallet/src/redux/channel/withdrawingReducer.ts
@@ -1,11 +1,11 @@
-import * as states from '../../channel/state';
-import * as actions from '../../actions';
-import { unreachable } from '../../../utils/reducer-utils';
-import { handleSignatureAndValidationMessages } from '../../../utils/state-utils';
-import { createTransferAndWithdrawTransaction } from '../../../utils/transaction-generator';
-import { signVerificationData } from '../../../utils/signing-utils';
+import * as states from './state';
+import * as actions from '../actions';
+import { unreachable } from '../../utils/reducer-utils';
+import { handleSignatureAndValidationMessages } from '../../utils/state-utils';
+import { createTransferAndWithdrawTransaction } from '../../utils/transaction-generator';
+import { signVerificationData } from '../../utils/signing-utils';
 import { closeSuccess, hideWallet } from 'magmo-wallet-client/lib/wallet-events';
-import { NextChannelState } from '../../sharedState';
+import { NextChannelState } from '../sharedState';
 
 export const withdrawingReducer = (
   state: states.WithdrawingState,
